renaming dialogueSuggestionsPage to dialogueMatchingPage
